Feat: MOE refactor part 1
