+++ src/prime_rl/trainer/model.py
@@ -356,21 +356,26 @@ def is_tt_moe_model(model: nn.Module) -> bool:
 
 
 def configure_moe_ep_backend(model: nn.Module, config: ModelConfig) -> None:
+    logger = get_logger()
     backend = config.ep_comm_backend
+    expert_backend = config.expert_backend
+
     if backend == "deepep":
         from prime_rl.trainer.distributed.deepep import configure_num_sms
 
         configure_num_sms(config.deepep_num_sms)
-        if config.deepep_expert_backend == "sonic" and "QUACK_COMPILE_WORKERS" not in os.environ:
-            os.environ["QUACK_COMPILE_WORKERS"] = "1"
-            get_logger().info("Set QUACK_COMPILE_WORKERS=1 for sonic MoE backend")
+
+    if expert_backend == "sonic" and "QUACK_COMPILE_WORKERS" not in os.environ:
+        os.environ["QUACK_COMPILE_WORKERS"] = "1"
+        logger.debug("Set QUACK_COMPILE_WORKERS=1 for sonic MoE backend")
+
     language_model = get_language_model(model)
     for transformer_block in language_model.layers:
         if not isinstance(transformer_block.mlp, (MoE, LatentMoE)):
             continue
         transformer_block.mlp.set_ep_comm_backend(backend)
-        if backend == "deepep":
-            transformer_block.mlp.set_deepep_expert_backend(config.deepep_expert_backend)
+
+        transformer_block.mlp.set_expert_backend(config.expert_backend)
 
 
 def get_load_balance_stats(
