+++ src/prime_rl/trainer/models/layers/checkpointing.py
@@ -110,9 +110,7 @@ def get_supported_targets(layer: nn.Module) -> frozenset[str]:
     if mlp is not None and _is_dense_mlp(mlp):
         supported_targets.add("mlp")
     if mlp is not None and (
-        hasattr(mlp, "_run_routed_experts")
-        or hasattr(mlp, "_run_local_routed_experts")
-        or hasattr(mlp, "_run_sonic_routed_experts")
+        hasattr(mlp, "run_experts_for_torch_a2a") or hasattr(mlp, "run_experts_for_deepep")
     ):
         supported_targets.add("routed_experts")
     if linear_attn is not None:
@@ -141,12 +139,10 @@ def set_selective_activation_checkpointing(layer: nn.Module, targets: Iterable[s
     if mlp is not None and "mlp" in enabled_targets:
         checkpoint_method(mlp, "forward")
     if mlp is not None and "routed_experts" in enabled_targets:
-        if hasattr(mlp, "_run_routed_experts"):
-            checkpoint_method(mlp, "_run_routed_experts")
-        if hasattr(mlp, "_run_local_routed_experts"):
-            checkpoint_method(mlp, "_run_local_routed_experts")
-        if hasattr(mlp, "_run_sonic_routed_experts"):
-            checkpoint_method(mlp, "_run_sonic_routed_experts")
+        if hasattr(mlp, "run_experts_for_torch_a2a"):
+            checkpoint_method(mlp, "run_experts_for_torch_a2a")
+        if hasattr(mlp, "run_experts_for_deepep"):
+            checkpoint_method(mlp, "run_experts_for_deepep")
     if linear_attn is not None and "linear_attn" in enabled_targets:
         checkpoint_method(linear_attn, "forward")
     if "norm" in enabled_targets:
