--- src/prime_rl/trainer/models/layers/moe.py
@@ -1,1007 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-from dataclasses import dataclass
-from functools import wraps
-from typing import Literal
-
-import torch
-import torch.nn.functional as F
-from torch import nn
-from torchtitan.distributed.expert_parallel import expert_parallel
-
-from prime_rl.configs.trainer import DeepEPExpertBackend, EPCommBackend
-
-
-_sonic_quack_autotune_disabled = False
-
-
-def _force_quack_default_config(fn):
-    if getattr(fn, "_prime_rl_force_quack_default_config", False):
-        return fn
-
-    @wraps(fn)
-    def wrapped(*args, **kwargs):
-        kwargs["tuned"] = False
-        return fn(*args, **kwargs)
-
-    wrapped._prime_rl_force_quack_default_config = True
-    return wrapped
-
-
-def _disable_sonic_quack_autotune() -> None:
-    global _sonic_quack_autotune_disabled
-    if _sonic_quack_autotune_disabled:
-        return
-
-    import sonicmoe.functional.backward as sonic_backward
-    import sonicmoe.functional.forward as sonic_forward
-
-    sonic_forward.gemm = _force_quack_default_config(sonic_forward.gemm)
-    sonic_forward.gemm_gated = _force_quack_default_config(sonic_forward.gemm_gated)
-    sonic_backward.gemm = _force_quack_default_config(sonic_backward.gemm)
-    sonic_backward.gemm_dgated = _force_quack_default_config(sonic_backward.gemm_dgated)
-    _sonic_quack_autotune_disabled = True
-
-
-@dataclass
-class MoEArgs:
-    num_experts: int = 8
-    num_shared_experts: int = 1
-
-    # router
-    score_func: Literal["softmax", "sigmoid"] = "sigmoid"
-    route_norm: bool = False
-    route_scale: float = 1.0
-    score_before_experts: bool = True
-
-    # token-choice
-    top_k: int = 1
-    use_grouped_mm: bool = True  # grouped mm or for-loop for the experts computation
-    load_balance_coeff: float | None = 1e-3
-
-    def __post_init__(self):
-        assert self.use_grouped_mm, "use_grouped_mm must be True"
-
-
-# can be used as dense FFN layer or shared experts in MoE layers
-class FeedForward(nn.Module):
-    """
-    Args:
-        dim (int): Input dimension.
-        hidden_dim (int): Hidden dimension of the feedforward layer.
-
-    Attributes:
-        w1 (Linear): Linear transformation for the first layer.
-        w2 (Linear): Linear transformation for the second layer.
-        w3 (Linear): Linear transformation for the third layer.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        hidden_dim: int,
-    ) -> None:
-        super().__init__()
-        self.w1 = nn.Linear(dim, hidden_dim, bias=False)
-        self.w2 = nn.Linear(hidden_dim, dim, bias=False)
-        self.w3 = nn.Linear(dim, hidden_dim, bias=False)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self.w2(F.silu(self.w1(x)) * self.w3(x))
-
-    def init_weights(self, init_std: float = 0.02):
-        nn.init.trunc_normal_(self.w1.weight, mean=0.0, std=0.02)
-        for linear in (self.w2, self.w3):
-            nn.init.trunc_normal_(linear.weight, mean=0.0, std=init_std)
-
-
-class BCFeedForward(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        hidden_dim: int,
-    ) -> None:
-        super().__init__()
-        self.w1 = nn.Parameter(torch.empty(hidden_dim, dim))
-        self.w2 = nn.Parameter(torch.empty(dim, hidden_dim))
-        self.w3 = nn.Parameter(torch.empty(hidden_dim, dim))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        # return torch.matmul(self.w2, F.silu(torch.matmul(self.w1, x.T)) * torch.matmul(self.w3, x.T))
-        return torch.matmul(F.silu(torch.matmul(x, self.w1.T)) * torch.matmul(x, self.w3.T), self.w2.T)
-
-    def init_weights(self, init_std: float):
-        nn.init.trunc_normal_(self.w1, mean=0.0, std=0.02)
-        nn.init.trunc_normal_(self.w2, mean=0.0, std=init_std)
-        nn.init.trunc_normal_(self.w3, mean=0.0, std=init_std)
-
-
-def _run_experts_grouped_mm_impl(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    w3: torch.Tensor,
-    x: torch.Tensor,
-    num_tokens_per_expert: torch.Tensor,
-) -> torch.Tensor:
-    offsets = torch.cumsum(num_tokens_per_expert, dim=0, dtype=torch.int32)
-    # grouped mm between a 2D tensor and a 3D tensor
-    assert x.dim() == 2
-
-    h = F.silu(torch._grouped_mm(x.bfloat16(), w1.bfloat16().transpose(-2, -1), offs=offsets))
-    h = h * torch._grouped_mm(x.bfloat16(), w3.bfloat16().transpose(-2, -1), offs=offsets)
-    out = torch._grouped_mm(h, w2.bfloat16().transpose(-2, -1), offs=offsets).type_as(x)
-
-    return out
-
-
-def _run_experts_sonic_impl(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    w3: torch.Tensor,
-    x: torch.Tensor,
-    top_scores: torch.Tensor,
-    token_indices: torch.Tensor,
-    expert_indices: torch.Tensor,
-) -> torch.Tensor:
-    if expert_indices.numel() == 0:
-        return x.new_zeros(x.shape)
-
-    from sonicmoe.enums import ActivationType
-    from sonicmoe.functional import moe_general_routing_inputs
-
-    _disable_sonic_quack_autotune()
-
-    compute_x = x.bfloat16()
-    gate_up_proj = torch.stack(
-        [
-            w1.bfloat16().transpose(-2, -1),
-            w3.bfloat16().transpose(-2, -1),
-        ],
-        dim=-1,
-    ).flatten(-2)
-    out, _ = moe_general_routing_inputs(
-        compute_x,
-        top_scores,
-        token_indices,
-        expert_indices,
-        gate_up_proj.permute(2, 1, 0),
-        None,
-        w2.bfloat16().permute(1, 2, 0),
-        None,
-        w1.shape[0],
-        torch.cuda.current_stream().cuda_stream,
-        ActivationType.SWIGLU,
-        False,
-        False,
-    )
-    return out.type_as(x)
-
-
-class GroupedExperts(nn.Module):
-    def __init__(
-        self,
-        dim: int,
-        hidden_dim: int,
-        num_experts: int,
-        use_grouped_mm: bool,
-    ):
-        super().__init__()
-        self.num_experts = num_experts
-        self.w1 = nn.Parameter(torch.empty(num_experts, hidden_dim, dim))
-        self.w2 = nn.Parameter(torch.empty(num_experts, dim, hidden_dim))
-        self.w3 = nn.Parameter(torch.empty(num_experts, hidden_dim, dim))
-        self.ep_comm_backend: EPCommBackend = "torch"
-        self._forward_fn = expert_parallel(_run_experts_grouped_mm_impl)
-
-    def set_ep_comm_backend(self, backend: EPCommBackend) -> None:
-        self.ep_comm_backend = backend
-        if self.ep_comm_backend == "deepep":
-            self._forward_fn = _run_experts_grouped_mm_impl
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor | None = None,
-        *,
-        top_scores: torch.Tensor | None = None,
-        token_indices: torch.Tensor | None = None,
-        expert_indices: torch.Tensor | None = None,
-    ) -> torch.Tensor:
-        if self.ep_comm_backend == "deepep":
-            w1 = self.w1.to_local()
-            w2 = self.w2.to_local()
-            w3 = self.w3.to_local()
-        else:
-            w1 = self.w1
-            w2 = self.w2
-            w3 = self.w3
-
-        if top_scores is not None or token_indices is not None or expert_indices is not None:
-            if self.ep_comm_backend != "deepep":
-                raise ValueError("Sonic expert compute is only supported with model.ep_comm_backend='deepep'.")
-            if top_scores is None or token_indices is None or expert_indices is None:
-                raise ValueError("Sonic expert compute requires top_scores, token_indices, and expert_indices.")
-            return _run_experts_sonic_impl(w1, w2, w3, x, top_scores, token_indices, expert_indices)
-
-        if num_tokens_per_expert is None:
-            raise ValueError("num_tokens_per_expert is required for grouped-mm expert compute.")
-        return self._forward_fn(w1, w2, w3, x, num_tokens_per_expert)
-
-    def init_weights(self, init_std: float):
-        nn.init.trunc_normal_(self.w1, mean=0.0, std=0.02)
-        nn.init.trunc_normal_(self.w2, mean=0.0, std=init_std)
-        nn.init.trunc_normal_(self.w3, mean=0.0, std=init_std)
-
-
-class TokenChoiceTopKRouter(nn.Module):
-    """This class implements token-choice routing. In token-choice top-K routing, each token is
-        routed to top K experts based on the router scores.
-
-    Args:
-        dim (int): Dimension of input tokens.
-        num_experts (int): Number of experts in each moe layer.
-        top_k (int): Number of experts each token will be routed to in token-choice routing.
-        score_func (Literal["softmax", "sigmoid"]): Whether to use sigmoid or softmax for router scores.
-        route_norm (bool): Whether to normalize the routing scores when using sigmoid.
-        route_scale (float): Scaling factor applied to the routing scores.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        num_experts: int,
-        top_k: int,
-        score_func: Literal["softmax", "sigmoid"],
-        route_norm: bool,
-        route_scale: float,
-    ):
-        super().__init__()
-        self.gate = nn.Linear(dim, num_experts, bias=False)
-        self.num_experts = num_experts
-        self.top_k = top_k
-        self.score_func = score_func
-        self.route_norm = route_norm
-        self.route_scale = route_scale
-
-    def forward(
-        self, x: torch.Tensor, expert_bias: torch.Tensor | None = None, routed_experts: torch.Tensor | None = None
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """
-        Args:
-            x (torch.Tensor): Input tensor with shape ``(bs*slen, dim)``.
-            expert_bias (torch.Tensor | None, optional): Optional bias tensor for experts with shape ``(num_experts,)``.
-                Used for load balancing. Defaults to None.
-            routed_experts (torch.Tensor | None, optional): Optional tensor with shape ``(bs * slen, top_k)``.
-
-        Returns:
-            tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-                - top_scores (torch.Tensor):
-                    Routing scores for selected experts with shape ``(bs*slen, top_k)``.
-                - selected_experts_indices (torch.Tensor):
-                    Expert indices selected for each token with shape ``(bs*slen, top_k)``.
-                - num_tokens_per_expert (torch.Tensor):
-                    Number of tokens assigned to each expert with shape ``(num_experts,)``.
-        """
-        # scores shape (bs*slen, num_experts)
-        assert routed_experts is None or routed_experts.shape[-1] == self.top_k, (
-            f"routed_experts shape: {routed_experts.shape}, top_k: {self.top_k}"
-        )
-        scores = self.gate(x)
-
-        # By default, sigmoid or softmax is performed in float32 to avoid loss explosion
-        if self.score_func == "sigmoid":
-            scores = torch.sigmoid(scores.to(torch.float32))
-        elif self.score_func == "softmax":
-            scores = F.softmax(scores.to(torch.float32), dim=1)
-        else:
-            raise NotImplementedError(f"Unknown score function {self.score_func}")
-
-        # top scores shape (bs*slen, top_k)
-        # NOTE: The expert_bias is only used for routing. The gating value
-        #       top_scores is still derived from the original scores.
-
-        if routed_experts is not None:
-            top_scores = scores.gather(dim=1, index=routed_experts)
-            selected_experts_indices = routed_experts
-        elif expert_bias is not None:
-            _, selected_experts_indices = torch.topk(scores + expert_bias, k=self.top_k, dim=1)
-            top_scores = scores.gather(dim=1, index=selected_experts_indices)
-        else:
-            top_scores, selected_experts_indices = torch.topk(scores, k=self.top_k, dim=1)
-
-        if self.route_norm:
-            denominator = top_scores.sum(dim=-1, keepdim=True) + 1e-20
-            top_scores = top_scores / denominator
-        top_scores = top_scores * self.route_scale
-
-        # group tokens together by expert indices from 0 to num_experts and pass that to experts forward
-        num_tokens_per_expert = torch.histc(
-            selected_experts_indices.reshape(-1),
-            bins=self.num_experts,
-            min=0,
-            max=self.num_experts,
-        )
-
-        return top_scores, selected_experts_indices, num_tokens_per_expert
-
-    def init_weights(self, init_std: float):
-        nn.init.trunc_normal_(self.gate.weight, mean=0.0, std=init_std)
-
-
-# NOTE: the reason we make this a stateless module is to support
-#       expert_tensor_parallel_degree=1 with consistent TP/EP APIs.
-class TokenReorderer(nn.Module):
-    """
-    This module reorders token indices to match the order of experts, enabling
-    efficient parallel processing of tokens by experts.
-
-    Args:
-        num_experts (int): Number of experts in the MoE layer.
-        top_k (int): Number of experts each token will be routed to.
-    """
-
-    def __init__(self, num_experts: int, top_k: int):
-        super().__init__()
-        self.num_experts = num_experts
-        self.top_k = top_k
-
-    def forward(
-        self,
-        top_scores: torch.Tensor,
-        selected_experts_indices: torch.Tensor,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        """
-        Reorders token indices to match the order of experts for MoE routing.
-
-        Args:
-            top_scores (torch.Tensor): Routing scores for selected experts,
-                shape (batch_size*seq_len, top_k)
-            selected_experts_indices (torch.Tensor): Expert indices selected for each token,
-                shape (batch_size*seq_len, top_k)
-
-        Returns:
-            tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-                - top_scores_experts_sorted: Scores reordered to match expert ordering
-                - token_indices_experts_sorted: Token indices reordered to match expert ordering
-                - num_tokens_per_expert: Number of tokens assigned to each expert
-        """
-        # group tokens together by expert indices from 0 to num_experts and pass that to experts forward
-        selected_experts_indices = selected_experts_indices.reshape(-1)
-        num_tokens_per_expert = torch.histc(
-            selected_experts_indices,
-            bins=self.num_experts,
-            min=0,
-            max=self.num_experts,
-        )
-
-        # Reorder the token indices to match the order of the experts
-        # token_indices_experts_sorted shape (bs*slen*top_k,)
-        token_indices_experts_sorted = torch.argsort(selected_experts_indices, stable=True)
-
-        top_scores_experts_sorted = top_scores.view(-1)[token_indices_experts_sorted]
-        token_indices_experts_sorted = token_indices_experts_sorted // self.top_k
-
-        return (
-            top_scores_experts_sorted,
-            token_indices_experts_sorted,
-            num_tokens_per_expert,
-        )
-
-
-class MoE(nn.Module):
-    def __init__(self, moe_args: MoEArgs, dim: int, hidden_dim: int):
-        super().__init__()
-
-        num_experts = moe_args.num_experts
-        self.experts = GroupedExperts(
-            dim=dim,
-            hidden_dim=hidden_dim,
-            num_experts=num_experts,
-            use_grouped_mm=moe_args.use_grouped_mm,
-        )
-        self.ep_comm_backend: EPCommBackend = "torch"
-        self.experts.set_ep_comm_backend(self.ep_comm_backend)
-        self.router = TokenChoiceTopKRouter(
-            dim=dim,
-            num_experts=num_experts,
-            top_k=moe_args.top_k,
-            score_func=moe_args.score_func,
-            route_norm=moe_args.route_norm,
-            route_scale=moe_args.route_scale,
-        )
-        self.reorderer = TokenReorderer(num_experts=num_experts, top_k=moe_args.top_k)
-        # TODO: Add the s back and use FF when the weights support it
-        self.shared_expert = (
-            BCFeedForward(dim=dim, hidden_dim=hidden_dim * moe_args.num_shared_experts)
-            if moe_args.num_shared_experts > 0
-            else None
-        )
-        self.score_before_experts = moe_args.score_before_experts
-        self.deepep_expert_backend: DeepEPExpertBackend = "grouped_mm"
-
-        # define fields for auxiliary-loss-free load balancing (https://arxiv.org/abs/2408.15664)
-        # NOTE: tokens_per_expert is accumulated in the model forward pass.
-        #       expert_bias is updated outside the model in an optimizer step pre hook
-        #       to work with gradient accumulation.
-        self.load_balance_coeff = moe_args.load_balance_coeff
-        if self.load_balance_coeff is not None:
-            assert self.load_balance_coeff > 0.0
-            self.register_buffer(
-                "expert_bias",
-                torch.zeros(num_experts, dtype=torch.float32),
-                persistent=True,
-            )
-        else:
-            self.expert_bias = None
-        # tokens_per_expert will be used to track expert usage and to update the expert bias for load balancing
-        self.register_buffer(
-            "tokens_per_expert",
-            torch.zeros(num_experts, dtype=torch.float32),
-            persistent=False,
-        )
-
-    def set_ep_comm_backend(self, backend: EPCommBackend) -> None:
-        self.ep_comm_backend = backend
-        self.experts.set_ep_comm_backend(backend)
-
-    def set_deepep_expert_backend(self, backend: DeepEPExpertBackend) -> None:
-        if backend == "sonic" and self.score_before_experts:
-            raise ValueError("model.deepep_expert_backend='sonic' requires MoE score_before_experts=False.")
-        self.deepep_expert_backend = backend
-
-    def _run_local_routed_experts(
-        self,
-        x: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor,
-    ) -> torch.Tensor:
-        return self.experts(x, num_tokens_per_expert)
-
-    def _run_routed_experts(
-        self,
-        x: torch.Tensor,
-        token_indices_experts_sorted: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor,
-        top_scores_experts_sorted: torch.Tensor,
-    ) -> torch.Tensor:
-        dim = x.shape[-1]
-        routed_indices = token_indices_experts_sorted.reshape(-1, 1).expand(-1, dim)
-        routed_input = torch.gather(x, dim=0, index=routed_indices)
-
-        if self.score_before_experts:
-            routed_input = (routed_input.to(torch.float32) * top_scores_experts_sorted.reshape(-1, 1)).to(x.dtype)
-
-        routed_output = self.experts(routed_input, num_tokens_per_expert)
-
-        if not self.score_before_experts:
-            routed_output = (routed_output.to(torch.float32) * top_scores_experts_sorted.reshape(-1, 1)).to(x.dtype)
-
-        return routed_output
-
-    def _run_sonic_routed_experts(
-        self,
-        x: torch.Tensor,
-        top_scores: torch.Tensor,
-        token_indices: torch.Tensor,
-        expert_indices: torch.Tensor,
-    ) -> torch.Tensor:
-        return self.experts(
-            x,
-            top_scores=top_scores,
-            token_indices=token_indices,
-            expert_indices=expert_indices,
-        )
-
-    def _run_deepep_routed_experts(
-        self,
-        x: torch.Tensor,
-        selected_experts_indices: torch.Tensor,
-        top_scores: torch.Tensor,
-    ) -> torch.Tensor:
-        from prime_rl.trainer.distributed.deepep import (
-            combine_tokens,
-            dispatch_tokens,
-            dispatch_tokens_for_sonic,
-        )
-        from prime_rl.trainer.distributed.expert_parallel import get_ep_group
-
-        if x.shape[0] == 0:
-            shared_output = self.shared_expert(x) if self.shared_expert is not None else None
-            return x.new_zeros(x.shape) if shared_output is None else shared_output
-
-        group = get_ep_group(self.experts)
-        if self.deepep_expert_backend == "sonic":
-            hidden_states, expert_scores, token_indices, expert_indices, dispatch_state = dispatch_tokens_for_sonic(
-                x,
-                selected_experts_indices,
-                top_scores,
-                num_experts=self.experts.num_experts,
-                group=group,
-            )
-            routed_output = self._run_sonic_routed_experts(
-                hidden_states,
-                expert_scores,
-                token_indices,
-                expert_indices,
-            )
-        else:
-            hidden_states, num_tokens_per_expert, dispatch_state = dispatch_tokens(
-                x,
-                selected_experts_indices,
-                top_scores,
-                num_experts=self.experts.num_experts,
-                group=group,
-                score_before_experts=self.score_before_experts,
-            )
-            routed_output = self._run_local_routed_experts(hidden_states, num_tokens_per_expert)
-        routed_output = combine_tokens(routed_output, dispatch_state)
-        if self.shared_expert is not None:
-            routed_output = routed_output + self.shared_expert(x)
-
-        return routed_output
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        routed_experts: torch.Tensor | None = None,
-    ) -> torch.Tensor:
-        """
-        Args:
-            x (torch.Tensor): Input tensor with shape ``(bs, slen, dim)``.
-            routed_experts (torch.Tensor | None, optional): Optional tensor with shape ``(bs, slen, top_k)``.
-
-        Returns:
-            out (torch.Tensor): Output tensor with shape ``(bs, slen, dim)``.
-        """
-        bs, slen, dim = x.shape
-        x = x.view(-1, dim)
-
-        if routed_experts is not None:
-            _, _, top_k = routed_experts.shape
-            routed_experts = routed_experts.reshape(
-                -1, top_k
-            )  # we have to reshape here because the original is non-contiguous
-
-        # top_scores and selected_experts_indices shape (bs*slen*top_k,)
-        # num_tokens_per_expert shape (num_experts,)
-        (
-            top_scores,
-            selected_experts_indices,
-            num_tokens_per_expert,
-        ) = self.router(x, self.expert_bias, routed_experts=routed_experts)
-
-        # tokens_per_expert will be used to update the expert bias for load balancing.
-        # and also to count the expert usage
-        # Full block checkpointing can double count tokens_per_expert because it reruns the router
-        # in backward. The selective MoE path avoids that by checkpointing only the
-        # routed expert compute below.
-        with torch.no_grad():
-            self.tokens_per_expert.add_(num_tokens_per_expert)
-
-        if self.ep_comm_backend == "deepep":
-            routed_output = self._run_deepep_routed_experts(x, selected_experts_indices, top_scores)
-            return routed_output.reshape(bs, slen, dim)
-
-        # top_scores and token_indices_experts_sorted shape (bs*slen*top_k,)
-        # num_tokens_per_expert shape (num_experts,)
-        # NOTE: the reason we need to compute num_tokens_per_expert again is:
-        #       1st computation in router is to update self.tokens_per_expert
-        #       which would be the same across all TP ranks.
-        #       2nd computation in reorderer is for the actual routing and experts computation
-        #       which would be sharded over TP ranks if expert_tensor_parallel_degree==1.
-        #       If tensor_paralllel_degree == expert_tensor_parallel_degree, they agree.
-        (
-            top_scores_experts_sorted,
-            token_indices_experts_sorted,
-            num_tokens_per_expert,
-        ) = self.reorderer(top_scores, selected_experts_indices)
-
-        routed_output = self._run_routed_experts(
-            x,
-            token_indices_experts_sorted,
-            num_tokens_per_expert,
-            top_scores_experts_sorted,
-        )
-        if self.shared_expert is not None:
-            out = self.shared_expert(x)
-        else:
-            out = torch.zeros_like(x)
-
-        routed_indices = token_indices_experts_sorted.reshape(-1, 1).expand(-1, dim)
-        out = out.scatter_add(dim=0, index=routed_indices, src=routed_output)
-        out = out.reshape(bs, slen, dim)
-        return out
-
-    def init_weights(
-        self,
-        init_std: float,
-        buffer_device: torch.device,
-    ):
-        self.experts.init_weights(init_std)
-        self.router.init_weights(init_std)
-        if self.shared_expert is not None:
-            self.shared_expert.init_weights(init_std)
-
-        with torch.device(buffer_device):
-            self.tokens_per_expert = torch.zeros(self.experts.num_experts, dtype=torch.float32)
-            if self.load_balance_coeff is not None:
-                self.expert_bias = torch.zeros(self.experts.num_experts, dtype=torch.float32)
-
-
-def relu2(x: torch.Tensor) -> torch.Tensor:
-    return F.relu(x).square()
-
-
-def _run_nongated_experts_for_loop_impl(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    _w3: torch.Tensor,
-    x: torch.Tensor,
-    num_tokens_per_expert: torch.Tensor,
-) -> torch.Tensor:
-    num_tokens_per_expert = num_tokens_per_expert.tolist()
-    num_padding = x.shape[0] - sum(num_tokens_per_expert)
-
-    x = torch.split(
-        x[: sum(num_tokens_per_expert)],
-        split_size_or_sections=num_tokens_per_expert,
-        dim=0,
-    )
-    out_experts_splits = []
-    for expert_idx, x_expert in enumerate(x):
-        h = relu2(torch.matmul(x_expert, w1[expert_idx].transpose(-2, -1)))
-        h = torch.matmul(h, w2[expert_idx].transpose(-2, -1))
-        out_experts_splits.append(h)
-    out = torch.cat(out_experts_splits, dim=0)
-    out = torch.vstack((out, out.new_zeros((num_padding, out.shape[-1]))))
-    return out
-
-
-@expert_parallel
-def _run_nongated_experts_for_loop(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    _w3: torch.Tensor,
-    x: torch.Tensor,
-    num_tokens_per_expert: torch.Tensor,
-) -> torch.Tensor:
-    return _run_nongated_experts_for_loop_impl(w1, w2, _w3, x, num_tokens_per_expert)
-
-
-def _run_nongated_experts_grouped_mm_impl(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    _w3: torch.Tensor,
-    x: torch.Tensor,
-    num_tokens_per_expert: torch.Tensor,
-) -> torch.Tensor:
-    offsets = torch.cumsum(num_tokens_per_expert, dim=0, dtype=torch.int32)
-    assert x.dim() == 2
-
-    h = relu2(torch._grouped_mm(x.bfloat16(), w1.bfloat16().transpose(-2, -1), offs=offsets))
-    out = torch._grouped_mm(h, w2.bfloat16().transpose(-2, -1), offs=offsets).type_as(x)
-    return out
-
-
-@expert_parallel
-def _run_nongated_experts_grouped_mm(
-    w1: torch.Tensor,
-    w2: torch.Tensor,
-    _w3: torch.Tensor,
-    x: torch.Tensor,
-    num_tokens_per_expert: torch.Tensor,
-) -> torch.Tensor:
-    return _run_nongated_experts_grouped_mm_impl(w1, w2, _w3, x, num_tokens_per_expert)
-
-
-class NonGatedGroupedExperts(nn.Module):
-    def __init__(
-        self,
-        input_dim: int,
-        intermediate_dim: int,
-        num_experts: int,
-        use_grouped_mm: bool,
-    ):
-        super().__init__()
-        self.num_experts = num_experts
-        self.w1 = nn.Parameter(torch.empty(num_experts, intermediate_dim, input_dim))
-        self.w2 = nn.Parameter(torch.empty(num_experts, input_dim, intermediate_dim))
-        # Dummy w3 for @expert_parallel decorator compatibility (expects w1, w2, w3 signature)
-        self.w3 = nn.Parameter(torch.empty(0))
-        self.use_grouped_mm = use_grouped_mm
-        self.ep_comm_backend: EPCommBackend = "torch"
-
-    def set_ep_comm_backend(self, backend: EPCommBackend) -> None:
-        self.ep_comm_backend = backend
-
-    def _forward_deepep(self, x: torch.Tensor, num_tokens_per_expert: torch.Tensor) -> torch.Tensor:
-        w1 = self.w1.to_local()
-        w2 = self.w2.to_local()
-        w3 = self.w3.to_local()
-        if self.use_grouped_mm:
-            return _run_nongated_experts_grouped_mm_impl(w1, w2, w3, x, num_tokens_per_expert)
-        return _run_nongated_experts_for_loop_impl(w1, w2, w3, x, num_tokens_per_expert)
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor,
-    ) -> torch.Tensor:
-        if self.ep_comm_backend == "deepep":
-            return self._forward_deepep(x, num_tokens_per_expert)
-        if self.use_grouped_mm:
-            return _run_nongated_experts_grouped_mm(self.w1, self.w2, self.w3, x, num_tokens_per_expert)
-        else:
-            return _run_nongated_experts_for_loop(self.w1, self.w2, self.w3, x, num_tokens_per_expert)
-
-    def init_weights(self, init_std: float):
-        nn.init.trunc_normal_(self.w1, mean=0.0, std=0.02)
-        nn.init.trunc_normal_(self.w2, mean=0.0, std=init_std)
-
-
-class NemotronHRouter(nn.Module):
-    """Sigmoid router with group-based expert selection and e_score_correction_bias.
-
-    Follows the DeepseekV3 routing pattern: sigmoid scoring, group-based top-k selection,
-    and bias correction for load balancing.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        num_experts: int,
-        top_k: int,
-        n_group: int,
-        topk_group: int,
-        norm_topk_prob: bool,
-    ):
-        super().__init__()
-        self.gate = nn.Parameter(torch.empty(num_experts, dim))
-        self.register_buffer("e_score_correction_bias", torch.zeros(num_experts))
-        self.num_experts = num_experts
-        self.top_k = top_k
-        self.n_group = n_group
-        self.topk_group = topk_group
-        self.norm_topk_prob = norm_topk_prob
-
-    def forward(
-        self, x: torch.Tensor, expert_bias: torch.Tensor | None = None
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        scores = F.linear(x.float(), self.gate.float()).sigmoid()
-        scores_for_choice = scores + self.e_score_correction_bias
-
-        if expert_bias is not None:
-            scores_for_choice = scores_for_choice + expert_bias
-
-        # Group-based routing
-        if self.n_group > 1:
-            group_scores = (
-                scores_for_choice.view(-1, self.n_group, self.num_experts // self.n_group)
-                .topk(2, dim=-1)[0]
-                .sum(dim=-1)
-            )
-            group_idx = torch.topk(group_scores, k=self.topk_group, dim=-1, sorted=False)[1]
-            group_mask = torch.zeros_like(group_scores)
-            group_mask.scatter_(1, group_idx, 1)
-            score_mask = (
-                group_mask.unsqueeze(-1)
-                .expand(-1, self.n_group, self.num_experts // self.n_group)
-                .reshape(-1, self.num_experts)
-            )
-            scores_for_choice = scores_for_choice.masked_fill(~score_mask.bool(), 0.0)
-
-        selected_experts_indices = torch.topk(scores_for_choice, k=self.top_k, dim=-1, sorted=False)[1]
-        top_scores = scores.gather(1, selected_experts_indices)
-
-        if self.norm_topk_prob:
-            denominator = top_scores.sum(dim=-1, keepdim=True) + 1e-20
-            top_scores = top_scores / denominator
-
-        num_tokens_per_expert = torch.histc(
-            selected_experts_indices.reshape(-1).float(),
-            bins=self.num_experts,
-            min=0,
-            max=self.num_experts,
-        )
-
-        return top_scores, selected_experts_indices, num_tokens_per_expert
-
-    def init_weights(self, init_std: float):
-        nn.init.trunc_normal_(self.gate, mean=0.0, std=init_std)
-
-
-class BCNonGatedFeedForward(nn.Module):
-    """Non-gated feed-forward network used as the shared expert in NemotronH.
-
-    Uses relu2 activation: down_proj(relu2(up_proj(x))).
-    """
-
-    def __init__(self, dim: int, hidden_dim: int):
-        super().__init__()
-        self.up_proj = nn.Linear(dim, hidden_dim, bias=False)
-        self.down_proj = nn.Linear(hidden_dim, dim, bias=False)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return self.down_proj(relu2(self.up_proj(x)))
-
-
-class LatentMoE(nn.Module):
-    """NemotronH-style Mixture of Experts with latent projections.
-
-    The input is projected to a latent space before expert computation,
-    and the output is projected back. Experts use relu2 activation without gating.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        latent_dim: int | None,
-        moe_intermediate_size: int,
-        shared_expert_intermediate_size: int,
-        num_experts: int,
-        top_k: int,
-        n_group: int,
-        topk_group: int,
-        norm_topk_prob: bool,
-        routed_scaling_factor: float,
-        use_grouped_mm: bool,
-        load_balance_coeff: float | None,
-    ):
-        super().__init__()
-        effective_latent_dim = latent_dim if latent_dim is not None else dim
-
-        self.router = NemotronHRouter(
-            dim=dim,
-            num_experts=num_experts,
-            top_k=top_k,
-            n_group=n_group,
-            topk_group=topk_group,
-            norm_topk_prob=norm_topk_prob,
-        )
-        self.experts = NonGatedGroupedExperts(
-            input_dim=effective_latent_dim,
-            intermediate_dim=moe_intermediate_size,
-            num_experts=num_experts,
-            use_grouped_mm=use_grouped_mm,
-        )
-        self.ep_comm_backend: EPCommBackend = "torch"
-        self.experts.set_ep_comm_backend(self.ep_comm_backend)
-        self.reorderer = TokenReorderer(num_experts=num_experts, top_k=top_k)
-        self.shared_expert = BCNonGatedFeedForward(dim=dim, hidden_dim=shared_expert_intermediate_size)
-
-        if latent_dim is not None:
-            self.fc1_latent_proj = nn.Linear(dim, latent_dim, bias=False)
-            self.fc2_latent_proj = nn.Linear(latent_dim, dim, bias=False)
-        else:
-            self.fc1_latent_proj = nn.Identity()
-            self.fc2_latent_proj = nn.Identity()
-
-        self.routed_scaling_factor = routed_scaling_factor
-        self.load_balance_coeff = load_balance_coeff
-        if self.load_balance_coeff is not None:
-            assert self.load_balance_coeff > 0.0
-            self.register_buffer(
-                "expert_bias",
-                torch.zeros(num_experts, dtype=torch.float32),
-                persistent=True,
-            )
-        else:
-            self.expert_bias = None
-        self.register_buffer(
-            "tokens_per_expert",
-            torch.zeros(num_experts, dtype=torch.float32),
-            persistent=False,
-        )
-
-    def set_ep_comm_backend(self, backend: EPCommBackend) -> None:
-        self.ep_comm_backend = backend
-        self.experts.set_ep_comm_backend(backend)
-
-    def set_deepep_expert_backend(self, backend: DeepEPExpertBackend) -> None:
-        if backend != "grouped_mm":
-            raise ValueError("LatentMoE only supports model.deepep_expert_backend='grouped_mm'.")
-
-    def _run_local_routed_experts(
-        self,
-        x: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor,
-    ) -> torch.Tensor:
-        return self.experts(x, num_tokens_per_expert)
-
-    def _run_routed_experts(
-        self,
-        x: torch.Tensor,
-        token_indices_experts_sorted: torch.Tensor,
-        num_tokens_per_expert: torch.Tensor,
-        top_scores_experts_sorted: torch.Tensor,
-    ) -> torch.Tensor:
-        dim = x.shape[-1]
-        token_indices_expanded = token_indices_experts_sorted.reshape(-1, 1).expand(-1, dim)
-        routed_input = torch.gather(x, dim=0, index=token_indices_expanded)
-
-        routed_input = self.fc1_latent_proj(routed_input)
-        routed_output = self.experts(routed_input, num_tokens_per_expert)
-
-        routed_output = (routed_output.float() * top_scores_experts_sorted.reshape(-1, 1)).to(routed_output.dtype)
-        routed_output = routed_output * self.routed_scaling_factor
-
-        routed_output = self.fc2_latent_proj(routed_output)
-        return routed_output
-
-    def _run_deepep_routed_experts(
-        self,
-        x: torch.Tensor,
-        selected_experts_indices: torch.Tensor,
-        top_scores: torch.Tensor,
-    ) -> torch.Tensor:
-        from prime_rl.trainer.distributed.deepep import (
-            combine_tokens,
-            dispatch_tokens,
-        )
-        from prime_rl.trainer.distributed.expert_parallel import get_ep_group
-
-        if x.shape[0] == 0:
-            return self.shared_expert(x)
-
-        group = get_ep_group(self.experts)
-        # Project before dispatch so DeepEP communicates the smaller latent activations.
-        latent_x = self.fc1_latent_proj(x)
-        hidden_states, num_tokens_per_expert, dispatch_state = dispatch_tokens(
-            latent_x,
-            selected_experts_indices,
-            top_scores,
-            num_experts=self.experts.num_experts,
-            group=group,
-            score_before_experts=False,
-        )
-        routed_output = self._run_local_routed_experts(hidden_states, num_tokens_per_expert)
-        routed_output = combine_tokens(routed_output, dispatch_state)
-        shared_output = self.shared_expert(x)
-        routed_output = routed_output * self.routed_scaling_factor
-        routed_output = self.fc2_latent_proj(routed_output)
-        return shared_output + routed_output
-
-    def forward(self, x: torch.Tensor, routed_experts: torch.Tensor | None = None) -> torch.Tensor:
-        bs, slen, dim = x.shape
-        x_flat = x.view(-1, dim)
-
-        top_scores, selected_experts_indices, num_tokens_per_expert = self.router(x_flat, self.expert_bias)
-
-        with torch.no_grad():
-            self.tokens_per_expert.add_(num_tokens_per_expert)
-
-        if self.ep_comm_backend == "deepep":
-            routed_output = self._run_deepep_routed_experts(x_flat, selected_experts_indices, top_scores)
-            return routed_output.reshape(bs, slen, dim)
-
-        (
-            top_scores_experts_sorted,
-            token_indices_experts_sorted,
-            num_tokens_per_expert,
-        ) = self.reorderer(top_scores, selected_experts_indices)
-
-        routed_output = self._run_routed_experts(
-            x_flat,
-            token_indices_experts_sorted,
-            num_tokens_per_expert,
-            top_scores_experts_sorted,
-        )
-
-        out = self.shared_expert(x_flat)
-
-        token_indices_full = token_indices_experts_sorted.reshape(-1, 1).expand(-1, dim)
-        out = out.scatter_add(dim=0, index=token_indices_full, src=routed_output)
-        out = out.reshape(bs, slen, dim)
-        return out
-
-    def init_weights(self, init_std: float, buffer_device: torch.device):
-        self.experts.init_weights(init_std)
-        self.router.init_weights(init_std)
-
-        with torch.device(buffer_device):
-            self.tokens_per_expert = torch.zeros(self.experts.num_experts, dtype=torch.float32)
-            if self.load_balance_coeff is not None:
-                self.expert_bias = torch.zeros(self.experts.num_experts, dtype=torch.float32)
