+++ src/prime_rl/trainer/models/layers/moe/__init__.py
@@ -1,0 +1,20 @@
+from prime_rl.trainer.models.layers.moe.experts import GroupedExperts, NonGatedGroupedExperts
+from prime_rl.trainer.models.layers.moe.ffn import BCFeedForward, BCNonGatedFeedForward, FeedForward
+from prime_rl.trainer.models.layers.moe.kernels import relu2
+from prime_rl.trainer.models.layers.moe.moe import LatentMoE, MoE, MoEArgs
+from prime_rl.trainer.models.layers.moe.routing import NemotronHRouter, TokenChoiceTopKRouter, TokenReorderer
+
+__all__ = [
+    "BCFeedForward",
+    "BCNonGatedFeedForward",
+    "FeedForward",
+    "GroupedExperts",
+    "LatentMoE",
+    "MoE",
+    "MoEArgs",
+    "NemotronHRouter",
+    "NonGatedGroupedExperts",
+    "TokenChoiceTopKRouter",
+    "TokenReorderer",
+    "relu2",
+]
